refactor(bench): split benchmark files per algorithm for granular CodSpeed reporting
Split index.bench.ts (mixed zstd+brotli) into separate zstd.bench.ts
and brotli.bench.ts so CodSpeed PR reports only show benchmarks relevant
to the changed algorithm. Extract shared test fixtures into
bench-fixtures.ts and add realistic data benchmarks (JSON, text) to
gzip and deflate for consistent coverage across all four algorithms.

Closes #173
