+++ __test__/deflate.bench.ts
@@ -1,28 +1,15 @@
 import { bench, describe } from 'vitest';
 import { deflateCompress, deflateDecompress } from '../index.js';
-
-// --- Patterned data (compressible) ---
-const SMALL = Buffer.from('Hello, zflate! '.repeat(10));
-const MEDIUM = Buffer.alloc(10_000);
-for (let i = 0; i < MEDIUM.length; i++) MEDIUM[i] = i % 256;
-const LARGE = Buffer.alloc(1_000_000);
-for (let i = 0; i < LARGE.length; i++) LARGE[i] = i % 256;
-
-// --- Deterministic pseudo-random data (incompressible) ---
-// Uses a linear congruential generator for reproducible benchmark inputs,
-// avoiding CodSpeed instruction count variance from non-deterministic randomBytes.
-const deterministicBytes = (size: number, seed: number): Buffer => {
-  const out = Buffer.alloc(size);
-  let x = seed >>> 0;
-  for (let i = 0; i < size; i++) {
-    x = (1664525 * x + 1013904223) >>> 0;
-    out[i] = x & 0xff;
-  }
-  return out;
-};
-const RANDOM_SMALL = deterministicBytes(150, 0x1234);
-const RANDOM_MEDIUM = deterministicBytes(10_000, 0x5678);
-const RANDOM_LARGE = deterministicBytes(1_000_000, 0x9abc);
+import {
+  JSON_DATA,
+  LARGE,
+  MEDIUM,
+  RANDOM_LARGE,
+  RANDOM_MEDIUM,
+  RANDOM_SMALL,
+  SMALL,
+  TEXT_DATA,
+} from './bench-fixtures.js';
 
 // --- Pre-compressed data for decompression benchmarks ---
 const SMALL_COMPRESSED = deflateCompress(SMALL);
@@ -31,6 +18,8 @@ const LARGE_COMPRESSED = deflateCompress(LARGE);
 const RANDOM_SMALL_COMPRESSED = deflateCompress(RANDOM_SMALL);
 const RANDOM_MEDIUM_COMPRESSED = deflateCompress(RANDOM_MEDIUM);
 const RANDOM_LARGE_COMPRESSED = deflateCompress(RANDOM_LARGE);
+const JSON_COMPRESSED = deflateCompress(JSON_DATA);
+const TEXT_COMPRESSED = deflateCompress(TEXT_DATA);
 
 describe('deflate compress (patterned)', () => {
   bench('150B', () => {
@@ -60,6 +49,16 @@ describe('deflate compress (random)', () => {
   });
 });
 
+describe('deflate compress (realistic)', () => {
+  bench(`JSON ${(JSON_DATA.length / 1024).toFixed(0)}KB`, () => {
+    deflateCompress(JSON_DATA);
+  });
+
+  bench(`text ${(TEXT_DATA.length / 1024).toFixed(0)}KB`, () => {
+    deflateCompress(TEXT_DATA);
+  });
+});
+
 describe('deflate decompress (patterned)', () => {
   bench('150B', () => {
     deflateDecompress(SMALL_COMPRESSED);
@@ -85,5 +84,15 @@ describe('deflate decompress (random)', () => {
 
   bench('1MB', () => {
     deflateDecompress(RANDOM_LARGE_COMPRESSED);
+  });
+});
+
+describe('deflate decompress (realistic)', () => {
+  bench(`JSON ${(JSON_DATA.length / 1024).toFixed(0)}KB`, () => {
+    deflateDecompress(JSON_COMPRESSED);
+  });
+
+  bench(`text ${(TEXT_DATA.length / 1024).toFixed(0)}KB`, () => {
+    deflateDecompress(TEXT_COMPRESSED);
   });
 });
